PIQL

Documentation for PIQL.

PIQL.ControlProblem

PIQL.PVActor

PIQL.PiqlParticle

PIQL.empty_actor

PIQL.initial_state_action

PIQL.qestimate

PIQL.run_piql!

PIQL.step_choices

PIQL.train!

PIQL.ControlProblem — Type

ControlProblem is a struct with fields that are the functions which completely define a KL-control problem

PIQL.PVActor — Type

We write this as a separation between policy and value function The policy is representable as a set of unnormalized weights w_i so that

log(pi) = log(wi) - log(sum(w)) ηi = log(wi/πi) - log(sum(w)) ηi = ωi - log(sum(πi * exp(ω_i)))

The ωi = log(wi/π_i) have an estimator in

β(criticq - actorv)i = hatωi => ω_i

There's no exponential sampling that needs to be averaged so we can simply update (with a pretty strong learning rate.)

ωi(t+1) = ωi(t) + α (hatωi - ωi(t))

at the same time, the actorq can be generated

actorqi = ηi / β + actorvi = (ωi - log(Ξ)) / β + actorv

and it can be constructed from

actorqi = ηi / β + actorvi = (qi - log(Ξ)) / β + actorv

v = sum(πi exp(β*actorqi)) / β w = β * (actorq_i - v)

Still experimental, loss interest after TabActor update rule didn't work.

PIQL.PiqlParticle — Type

current StateAction definition:

struct StateAction{S,A} # static and constructed on forward pass # atomic unit of data for all reinforcement learning # includes all information and diagnostics available from a single step. state::S action::A β::Float64 # the beta under which the temperature is allowed to fluxuate actorq::Float64 criticq::Float64 reward::Float64 # actually incurred cost entering the state V::Float64 # free energy of current action (observed value) U::Float64 # average energy of current action end

PIQL.empty_actor — Method

We define an actor interface: an actor is a callable struct

struct GenericActor parameters

end

function (a::Actor)(state,action) ... return E_actor end

function loss(actor, minibatch) ... return E_actor end

PIQL.initial_state_action — Method

Start off a trajectory with a new state action pair

PIQL.qestimate — Method

z(t) = exp(β(criticq - actorq)) (γ z(t+1) + (1-γ)) z(t) = exp(β(criticq - actorq)) (γ (z(t+1)-1) + 1) logz(t) = β(criticq - actorq) + log(γ(z(t+1)-1) + 1) = β(criticq - actorq) + logp1(γexpm1(logz(t+1)))

PIQL.run_piql! — Method

Expected dynamics: grow the worldline until the terminal condition. then backpropogate weights and return terminated = true

PIQL.step_choices — Method

Actions are indexed by natural numbers

PIQL.train! — Method

Destructive memory training, storing partition function memory is composed of struct (at time of writing) consisting of

struct QEstimate{S,A}
    sa::StateAction{S,A} # the state action information at this time step
    criticq::Float64 # from the state one step forward in time
    logz::Float64 # log z at state in sa.state
end